Added docs for other HTTP errors
This is needed when a user session expires (or other HTTP errors are thrown) in Laravel. Unpoly will swap the entire `<body>` of the document when an HTTP error occurs with the Laravel response, which will be a JSON "unauthenticated" message. I've looked at other ways of extending the `Illuminate\Http\Request`, but this appears to be the only way.

After this change though, Laravel will properly return a redirect response to the login page, which Unpoly will follow and grab the response from, displaying the login form and allowing users to login after their session has expired 👍